+++ server/src/lib/cloudinary.ts
@@ -1,0 +1,26 @@
+import {v2 as cloudinary} from "cloudinary";
+import {ENV} from "./env.js";
+
+// Cloudinary config
+cloudinary.config({
+  cloud_name: ENV.CLOUDINARY_CLOUD_NAME,
+  api_key: ENV.CLOUDINARY_API_KEY,
+  api_secret: ENV.CLOUDINARY_API_SECRET,
+  secure: true,
+});
+
+export const uploadFileToCloudinary = async (
+  fileBuffer: any,
+  folder: string
+) => {
+  return new Promise((resolve, reject) => {
+    const upload = cloudinary.uploader.upload_stream(
+      {folder},
+      (error, result) => {
+        if (error) return reject(error);
+        resolve(result);
+      }
+    );
+    upload.end(fileBuffer);
+  });
+};
