+++ server/src/lib/multer.ts
@@ -1,0 +1,6 @@
+import multer from "multer";
+
+const storage = multer.memoryStorage();
+const upload = multer({storage});
+
+export {upload};
