+++ server/src/server.ts
@@ -12,6 +12,7 @@ const startServer = async () => {
     });
   } catch (error) {
     console.error("Error starting server: ", error);
+    process.exit(1);
   }
 };
